set timestamp on creation recording of data callback message

+++ pkg/sigfox/backend/callback/data/message/message.go
@@ -2,7 +2,7 @@ package message
 
 type Message struct {
 	Id        string `json:"id" bson:"id"`
-	Timestamp int    `json:"timeStamp" bson:"timeStamp"`
+	Timestamp int64  `json:"timeStamp" bson:"timeStamp"`
 	DeviceId  string `json:"deviceId" bson:"deviceId"`
 	Data      []byte `json:"data" bson:"data"`
 }
